% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_dinamica_acoplamiento_adtr.R
\name{actualizar_escala_ecosistemica_nivel5_adtr}
\alias{actualizar_escala_ecosistemica_nivel5_adtr}
\title{Actualizar escala ecosistemica con Nivel 5 ADTR}
\usage{
actualizar_escala_ecosistemica_nivel5_adtr(evidencia_nivel5)
}
\arguments{
\item{evidencia_nivel5}{Resultado de [construir_evidencia_acoplamiento_nivel5_adtr()].}
}
\value{
data.frame con escala relacional actualizada.
}
\description{
Actualiza la escala relacional incorporando evidencia exploratoria de
acoplamiento interpretado.
}
