% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_reportes_adtr.R
\name{adtr_extraer}
\alias{adtr_extraer}
\title{Obtener un valor seguro desde una lista ADTR}
\usage{
adtr_extraer(x, nombre, default = NULL)
}
\arguments{
\item{x}{Lista.}

\item{nombre}{Nombre del componente.}

\item{default}{Valor por defecto.}
}
\value{
Componente solicitado o valor por defecto.
}
\description{
Función auxiliar interna para extraer componentes de una lista sin producir
errores si el componente no existe.
}
\keyword{internal}
