% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_metricas_adtr.R
\name{integrar_metricas_adtr}
\alias{integrar_metricas_adtr}
\title{Integrar métricas referenciales y transformacionales ADTR}
\usage{
integrar_metricas_adtr(metricas_referenciales, metricas_transformacionales)
}
\arguments{
\item{metricas_referenciales}{data.frame generado por
`calcular_metricas_referenciales_adtr()`.}

\item{metricas_transformacionales}{data.frame generado por
`calcular_metricas_transformacionales_adtr()`.}
}
\value{
data.frame con métricas integradas por punto.
}
\description{
Une las métricas provenientes de M(p) y K(p) en una sola tabla integrada.
}
