% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_dinamica_acoplamiento_adtr.R
\name{interpretar_acoplamiento_adtr}
\alias{interpretar_acoplamiento_adtr}
\title{Interpretar acoplamiento ADTR}
\usage{
interpretar_acoplamiento_adtr(
  regla_funcional,
  consistencia,
  dependencia,
  evidencia_nivel5,
  escala_actualizada
)
}
\arguments{
\item{regla_funcional}{Resultado de regla funcional.}

\item{consistencia}{Resultado de consistencia conjunta.}

\item{dependencia}{Resultado de dependencia exploratoria.}

\item{evidencia_nivel5}{Resultado de evidencia Nivel 5.}

\item{escala_actualizada}{Escala relacional actualizada.}
}
\value{
data.frame con interpretaciones.
}
\description{
Genera una interpretacion textual del acoplamiento interpretado exploratorio.
}
