% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_matrices_adtr.R
\name{unir_matrices_adtr}
\alias{unir_matrices_adtr}
\title{Unir matrices ADTR}
\usage{
unir_matrices_adtr(lista_matrices)
}
\arguments{
\item{lista_matrices}{Lista de data frames ADTR.}
}
\value{
Un `data.frame` con todas las matrices unidas.
}
\description{
Une una lista de matrices ADTR en una sola colección tabular. Todas las
matrices deben cumplir la estructura mínima ADTR.
}
